Create 0x29a99c126596c0Dc96b02A88a9EAab44EcCf511e.json
Cobalt – A new way to borrow!
The Cobalt lending platform aims to bring together borrowers into a fair democratic voting community in order to get all size loans from micro loans to six-figure loans all without the over sight of a bank, completely governed by the community and therefor the community directly profits from its’ involvement. The current credit system is fundamentally flawed and the Cobalt platform corrects this by facilitating a community voted and approved lending program on the blockchain that is completely based off the merits of the project, never seeing race, color, or individuals background. The only criteria considered for the loans will be merit of project and collateral provided by the borrower, then the community will be left to decide on funding the proposal or rejecting it.